requirements.txt
autopep8 == 1.2.4
Django == 1.9.5
django-bootstrap3 == 7.0.1
django-safedelete == 0.3.2
pep8 == 1.7.0
python-dateutil == 2.5.2
pytz == 2016.3
requests == 2.9.1
six == 1.10.0
